Create new file for attendees ui. Move and re-factor view_all_attendees
