+++ config/environment.rb
@@ -9,7 +9,8 @@ require "bundler/setup"
 Bundler.require(:default, ENV.fetch("RACK_ENV", nil))
 
 # Set up database connection
-set :database, { adapter: "sqlite3", database: "db/#{ENV.fetch('RACK_ENV', 'development')}.sqlite3" }
+set :database,
+    { adapter: "sqlite3", database: "db/#{ENV.fetch('RACK_ENV', 'development')}.sqlite3" }
 
 # Require in all files in 'app' directory
 require_all "app"
